changes to comms interfaces

+++ simbeeotic-core/src/main/java/harvard/robobees/simbeeotic/model/InertHive.java
@@ -35,7 +35,7 @@ public class InertHive implements GenericHiveLogic {
      * Does nothing.
      */
     @Override
-    public void messageReceived(double time, byte[] data, float rxPower) {
+    public void messageReceived(double time, byte[] data, double rxPower) {
     }
