+++ simbeeotic-examples/src/main/java/harvard/robobees/simbeeotic/example/RandomWalkBee.java
@@ -86,7 +86,7 @@ public class RandomWalkBee implements GenericBeeLogic {
 
 
     @Override
-    public void messageReceived(double time, byte[] data, float rxPower) {
+    public void messageReceived(double time, byte[] data, double rxPower) {
     }
